Refactor so more styling is shared
Between SingleChart and the Multi/Functions chart
