+++ src/charts/components/ErrorMessage/ErrorIcon.tsx
