+++ src/charts/components/Loading/Logo.tsx
